Moving linechart's axes to internal subdirectory.
