+++ widgets/linechart/internal/axes/label.go
