+++ widgets/linechart/internal/axes/scale.go
