+++ widgets/linechart/internal/axes/scale_test.go
