+++ widgets/linechart/linechart.go
@@ -30,7 +30,7 @@ import (
 	"github.com/mum4k/termdash/internal/numbers"
 	"github.com/mum4k/termdash/internal/terminalapi"
 	"github.com/mum4k/termdash/internal/widgetapi"
-	"github.com/mum4k/termdash/widgets/linechart/axes"
+	"github.com/mum4k/termdash/widgets/linechart/internal/axes"
 	"github.com/mum4k/termdash/widgets/linechart/zoom"
 )
