+++ widgets/linechart/options.go
@@ -19,7 +19,7 @@ import (
 	"math"
 
 	"github.com/mum4k/termdash/internal/cell"
-	"github.com/mum4k/termdash/widgets/linechart/axes"
+	"github.com/mum4k/termdash/widgets/linechart/internal/axes"
 	"github.com/mum4k/termdash/widgets/linechart/zoom"
 )
